+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- edctools.moveTabFirst
+- edctools.moveTabLast
+
 ## [0.0.2] - 2025-10-17
 
 ### Fixed